+++ projects/ngx-hateoas-client/src/lib/config/hateoas-configuration.interface.ts
@@ -62,6 +62,24 @@ export interface HateoasConfiguration {
    */
   isProduction?: boolean;
 
+  /**
+   * Specifying format for some type values.
+   */
+  typesFormat?: {
+    /**
+     * This date format will be used when parse {@link Resource} properties.
+     * If the property will be match to some one of specified formats, then the property type will be as Date object.
+     * Otherwise, raw type will be used as default.
+     */
+    date?: {
+      /**
+       * Date pattern.
+       * The {@link https://date-fns.org} lib is used to parse date with patterns, use patterns supported by this lib.
+       * See more about supported formats <a href='https://date-fns.org/v2.28.0/docs/parse'>here</a>.
+       */
+      patterns: Array<string>;
+    }
+  };
 
   /**
    * Let to change default page params that is size = 20, page = 0.